+++ CHANGELOG.md
@@ -5,6 +5,8 @@ v0.4.6
 ------
 
 * Added navigation option
+* Made left menu scrollable in lagoon template
+* Bug fixes
 
 v0.4.5
 ------